requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
cohere==5.15.0
python-dotenv==1.0.1
langdetect==1.0.9